FIX: make sure dryRun works

+++ src/Tasks/DeleteFilesFromDiskNotFoundInDatabase.php
@@ -22,20 +22,29 @@ class DeleteFilesFromDiskNotFoundInDatabase extends BuildTask
 
     public function run($request)
     {
-        $this->dryRun = $request->getVar('forreal') ? false : true;
+        $this->dryRun = $request && $request->getVar('forreal') ? false : true;
         $this->dryRunMessage();
         $files = AllFilesInfo::inst()->getAllFiles();
+        DB::alteration_message('=== START ===');
         foreach ($files as $file) {
             $array = $file->toMap();
             if (!empty($array['ErrorDBNotPresent'])) {
                 DB::alteration_message('--- Deleting from DB' . $array['Path'], 'deleted');
-                try {
-                    unlink($array['AbsolutePath']);
-                } catch (Exception $e) {
-                    DB::alteration_message('Error: ' . $e->getMessage(), 'deleted');
+                if ($this->dryRun) {
+                    DB::alteration_message('--- --- DRY RUN ONLY ---', 'deleted');
+                } else {
+                    try {
+                        unlink($array['AbsolutePath']);
+                    } catch (Exception $e) {
+                        DB::alteration_message('Error: ' . $e->getMessage(), 'deleted');
+                    }
                 }
+            } else {
+                echo '✓';
             }
         }
+        echo PHP_EOL;
+        DB::alteration_message('=== END ===');
 
         $this->dryRunMessage();
         DB::alteration_message('=== DONE ===', '');
